frontend most of the stuff
still missing settings

note: I modified a line in package.json. I had python 3.12 installed so I couldn't run it, and the line in setup-python assumes you already have python 3.11 as default, so I switched it to py - 3.11 so that it uses the py launcher and selects 3.11 manually. if it doesn't work for you just switch it back. I'll just keep the change for my device only so that I can run it